+++ packages/core/src/contexts/DriverContext.ts
@@ -10,6 +10,7 @@ import { Arguments, Argv } from '../types';
 
 export interface DriverArgs {
   concurrency: number;
+  live: boolean;
   name: string;
   priority: boolean;
   workspaces: string;